+++ kwcoco/util/util_special_json.py
@@ -1,7 +1,7 @@
 """
 Special non-general json functions
 """
-import ubelt as ub
+# import ubelt as ub
 from packaging.version import parse as Version
 import os
 import json as pjson